feat refactor - client opens - thumbnail list added to store
