+++ src/model/SocketIoConstants.ts
@@ -1,0 +1,2 @@
+export const IO_TIME_UPDATE = 'timeUpdated'
+export const IO_MEDIA_UPDATE = 'mediaUpdate'
